refactor: separate compoent for all forms
